refactor: integrate src/types/ into src/domain/
