+++ src/domain/AssetConfiguration.ts
