+++ src/domain/DynamicAssetConfiguration.ts
