+++ src/domain/EntityStateFlags.ts
