+++ src/domain/EventPriority.ts
