+++ src/domain/Game.ts
@@ -1,38 +1,38 @@
 import * as pl from "@akashic/playlog";
 import { Trigger } from "@akashic/trigger";
-import { AssetManager } from "./domain/AssetManager";
-import { AudioSystemManager } from "./domain/AudioSystemManager";
-import { Camera } from "./domain/Camera";
-import { DefaultLoadingScene } from "./domain/DefaultLoadingScene";
-import { E, PointSource } from "./domain/entities/E";
-import { Event, EventTypeString, JoinEvent, LeaveEvent, SeedEvent, PlayerInfoEvent } from "./domain/Event";
-import { EventConverter } from "./domain/EventConverter";
-import { LoadingScene } from "./domain/LoadingScene";
-import { ModuleManager } from "./domain/ModuleManager";
-import { OperationPluginManager } from "./domain/OperationPluginManager";
-import { PointEventResolver } from "./domain/PointEventResolver";
-import { RandomGenerator } from "./domain/RandomGenerator";
-import { Storage } from "./domain/Storage";
-import { SurfaceAtlasSet } from "./domain/SurfaceAtlasSet";
-import { SurfaceAtlasSetLike } from "./domain/SurfaceAtlasSetLike";
-import { XorshiftRandomGenerator } from "./domain/XorshiftRandomGenerator";
+import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
+import { AssetLike } from "../pdi-types/AssetLike";
+import { CommonOffset, CommonSize } from "../pdi-types/commons";
+import { PlatformPointEvent, PlatformPointType } from "../pdi-types/PlatformPointEvent";
+import { RendererLike } from "../pdi-types/RendererLike";
+import { ResourceFactoryLike } from "../pdi-types/ResourceFactoryLike";
+import { ScriptAssetRuntimeValueBase } from "../pdi-types/ScriptAssetRuntimeValue";
+import { AssetManager } from "./AssetManager";
+import { AudioSystemManager } from "./AudioSystemManager";
+import { Camera } from "./Camera";
+import { DefaultLoadingScene } from "./DefaultLoadingScene";
+import { E, PointSource } from "./entities/E";
+import { Event, EventTypeString, JoinEvent, LeaveEvent, SeedEvent, PlayerInfoEvent } from "./Event";
+import { EventConverter } from "./EventConverter";
+import { EventFilter } from "./EventFilter";
+import { GameConfiguration } from "./GameConfiguration";
 import { GameHandlerSet } from "./GameHandlerSet";
-import { ExceptionFactory } from "./pdi-common-impls/ExceptionFactory";
-import { AssetLike } from "./pdi-types/AssetLike";
-import { CommonOffset, CommonSize } from "./pdi-types/commons";
-import { PlatformPointEvent, PlatformPointType } from "./pdi-types/PlatformPointEvent";
-import { RendererLike } from "./pdi-types/RendererLike";
-import { ResourceFactoryLike } from "./pdi-types/ResourceFactoryLike";
-import { ScriptAssetRuntimeValueBase } from "./pdi-types/ScriptAssetRuntimeValue";
+import { GameMainParameterObject } from "./GameMainParameterObject";
+import { LoadingScene } from "./LoadingScene";
+import { LocalTickModeString } from "./LocalTickModeString";
+import { ModuleManager } from "./ModuleManager";
+import { OperationPlugin } from "./OperationPlugin";
+import { OperationPluginManager } from "./OperationPluginManager";
+import { InternalOperationPluginOperation } from "./OperationPluginOperation";
+import { OperationPluginViewInfo } from "./OperationPluginViewInfo";
+import { PointEventResolver } from "./PointEventResolver";
+import { RandomGenerator } from "./RandomGenerator";
 import { Scene } from "./Scene";
-import { EventFilter } from "./types/EventFilter";
-import { GameConfiguration } from "./types/GameConfiguration";
-import { GameMainParameterObject } from "./types/GameMainParameterObject";
-import { LocalTickModeString } from "./types/LocalTickModeString";
-import { OperationPlugin } from "./types/OperationPlugin";
-import { InternalOperationPluginOperation } from "./types/OperationPluginOperation";
-import { OperationPluginViewInfo } from "./types/OperationPluginViewInfo";
-import { TickGenerationModeString } from "./types/TickGenerationModeString";
+import { Storage } from "./Storage";
+import { SurfaceAtlasSet } from "./SurfaceAtlasSet";
+import { SurfaceAtlasSetLike } from "./SurfaceAtlasSetLike";
+import { TickGenerationModeString } from "./TickGenerationModeString";
+import { XorshiftRandomGenerator } from "./XorshiftRandomGenerator";
 
 /**
  * post-tick タスクのタイプ。
