+++ src/domain/GameConfiguration.ts
