+++ src/domain/GameMainParameterObject.ts
