+++ src/domain/LoadingScene.ts
@@ -1,7 +1,7 @@
 import { Trigger } from "@akashic/trigger";
 import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
 import { AssetLike } from "../pdi-types/AssetLike";
-import { Scene, SceneParameterObject } from "../Scene";
+import { Scene, SceneParameterObject } from "./Scene";
 
 export interface LoadingSceneParameterObject extends SceneParameterObject {
 	/**
