+++ src/domain/LocalTickModeString.ts
