+++ src/domain/OperationPluginInfo.ts
