+++ src/domain/OperationPluginManager.ts
@@ -1,10 +1,10 @@
 import { Trigger } from "@akashic/trigger";
-import { Game } from "../Game";
-import { OperationPlugin } from "../types/OperationPlugin";
-import { InternalOperationPluginInfo } from "../types/OperationPluginInfo";
-import { InternalOperationPluginOperation, OperationPluginOperation } from "../types/OperationPluginOperation";
-import { OperationPluginStatic } from "../types/OperationPluginStatic";
-import { OperationPluginViewInfo } from "../types/OperationPluginViewInfo";
+import { Game } from "./Game";
+import { OperationPlugin } from "./OperationPlugin";
+import { InternalOperationPluginInfo } from "./OperationPluginInfo";
+import { InternalOperationPluginOperation, OperationPluginOperation } from "./OperationPluginOperation";
+import { OperationPluginStatic } from "./OperationPluginStatic";
+import { OperationPluginViewInfo } from "./OperationPluginViewInfo";
 
 /**
  * 操作プラグインからの通知をハンドルするクラス。
