+++ src/domain/OperationPluginStatic.ts
