+++ src/domain/OperationPluginView.ts
