+++ src/domain/OperationPluginViewInfo.ts
