+++ src/domain/Scene.ts
@@ -1,24 +1,24 @@
 import { Trigger } from "@akashic/trigger";
+import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
+import { AssetLike } from "../pdi-types/AssetLike";
+import { AssetLoadFailureInfo } from "../pdi-types/AssetLoadFailureInfo";
+import { CommonOffset } from "../pdi-types/commons";
+import { StorageLoadError } from "../pdi-types/errors";
+import { AssetAccessor } from "./AssetAccessor";
 import { AssetHolder } from "./AssetHolder";
-import { AssetAccessor } from "./domain/AssetAccessor";
-import { Camera } from "./domain/Camera";
-import { Camera2D } from "./domain/Camera2D";
-import { E, PointDownEvent, PointMoveEvent, PointSource, PointUpEvent } from "./domain/entities/E";
-import { MessageEvent, OperationEvent } from "./domain/Event";
-import { getGameInAssetContext } from "./domain/getGameInAssetContext";
-import { Matrix } from "./domain/Matrix";
-import { StorageLoader, StorageLoaderHandler, StorageReadKey, StorageValueStore, StorageValueStoreSerialization } from "./domain/Storage";
-import { Timer } from "./domain/Timer";
-import { TimerIdentifier, TimerManager } from "./domain/TimerManager";
+import { Camera } from "./Camera";
+import { Camera2D } from "./Camera2D";
+import { DynamicAssetConfiguration } from "./DynamicAssetConfiguration";
+import { E, PointDownEvent, PointMoveEvent, PointSource, PointUpEvent } from "./entities/E";
+import { MessageEvent, OperationEvent } from "./Event";
 import { Game } from "./Game";
-import { ExceptionFactory } from "./pdi-common-impls/ExceptionFactory";
-import { AssetLike } from "./pdi-types/AssetLike";
-import { AssetLoadFailureInfo } from "./pdi-types/AssetLoadFailureInfo";
-import { CommonOffset } from "./pdi-types/commons";
-import { StorageLoadError } from "./pdi-types/errors";
-import { DynamicAssetConfiguration } from "./types/DynamicAssetConfiguration";
-import { LocalTickModeString } from "./types/LocalTickModeString";
-import { TickGenerationModeString } from "./types/TickGenerationModeString";
+import { getGameInAssetContext } from "./getGameInAssetContext";
+import { LocalTickModeString } from "./LocalTickModeString";
+import { Matrix } from "./Matrix";
+import { StorageLoader, StorageLoaderHandler, StorageReadKey, StorageValueStore, StorageValueStoreSerialization } from "./Storage";
+import { TickGenerationModeString } from "./TickGenerationModeString";
+import { Timer } from "./Timer";
+import { TimerIdentifier, TimerManager } from "./TimerManager";
 
 export type SceneRequestAssetHandler = () => void;
