+++ src/domain/TextAlign.ts
