+++ src/domain/TextAlignString.ts
