+++ src/domain/TextMetrics.ts
