+++ src/domain/TickGenerationModeString.ts
