+++ src/domain/entities/E.ts
@@ -8,15 +8,15 @@ import {
 	PointUpEventBase
 } from "../../domain/Event";
 import { Matrix, PlainMatrix } from "../../domain/Matrix";
-import { Game } from "../../Game";
 import { ExceptionFactory } from "../../pdi-common-impls/ExceptionFactory";
 import { CommonArea, CommonOffset, CommonRect } from "../../pdi-types/commons";
 import { RendererLike } from "../../pdi-types/RendererLike";
 import { ShaderProgramLike } from "../../pdi-types/ShaderProgramLike";
-import { Scene } from "../../Scene";
-import { EntityStateFlags } from "../../types/EntityStateFlags";
 import { Camera } from "../Camera";
+import { EntityStateFlags } from "../EntityStateFlags";
+import { Game } from "../Game";
 import { Object2D, Object2DParameterObject } from "../Object2D";
+import { Scene } from "../Scene";
 import { Util } from "../Util";
 
 /**
