+++ src/domain/entities/Label.ts
@@ -1,8 +1,8 @@
 import { Font } from "../../domain/Font";
 import { GlyphLike } from "../../pdi-types/GlyphLike";
 import { RendererLike } from "../../pdi-types/RendererLike";
-import { TextAlign } from "../../types/TextAlign";
-import { TextAlignString } from "../../types/TextAlignString";
+import { TextAlign } from "../TextAlign";
+import { TextAlignString } from "../TextAlignString";
 import { Util } from "../Util";
 import { CacheableE, CacheableEParameterObject } from "./CacheableE";
