+++ src/domain/getGameInAssetContext.ts
@@ -1,5 +1,5 @@
-import { Game } from "../Game";
 import { ExceptionFactory } from "../pdi-common-impls/ExceptionFactory";
+import { Game } from "./Game";
 
 declare const g: { game: Game };
