+++ src/index.common.ts
@@ -1,7 +1,7 @@
 export * from "@akashic/trigger";
 
-export * from "./Scene";
-export * from "./Game";
+export * from "./domain/Scene";
+export * from "./domain/Game";
 
 export * from "./pdi-common-impls/ExceptionFactory";
 export * from "./pdi-common-impls/Glyph";
@@ -93,35 +93,35 @@ export * from "./pdi-types/VideoSystemLike";
 
 export * from "./pdi-types/AudioAssetHint";
 export * from "./pdi-types/ImageAssetHint";
-export * from "./types/AssetConfiguration";
+export * from "./domain/AssetConfiguration";
 export * from "./pdi-types/AssetLoadErrorType";
 export * from "./pdi-types/commons";
 export * from "./pdi-types/CompositeOperation";
 export * from "./pdi-types/CompositeOperationString";
-export * from "./types/DynamicAssetConfiguration";
-export * from "./types/EntityStateFlags";
+export * from "./domain/DynamicAssetConfiguration";
+export * from "./domain/EntityStateFlags";
 export * from "./pdi-types/errors";
-export * from "./types/EventFilter";
-export * from "./types/EventIndex";
-export * from "./types/EventPriority";
+export * from "./domain/EventFilter";
+export * from "./domain/EventIndex";
+export * from "./domain/EventPriority";
 export * from "./pdi-types/FontFamily";
 export * from "./pdi-types/FontWeight";
 export * from "./pdi-types/FontWeightString";
-export * from "./types/GameConfiguration";
-export * from "./types/GameMainParameterObject";
+export * from "./domain/GameConfiguration";
+export * from "./domain/GameMainParameterObject";
 export * from "./pdi-types/ImageData";
-export * from "./types/LocalTickModeString";
-export * from "./types/OperationPlugin";
-export * from "./types/OperationPluginInfo";
-export * from "./types/OperationPluginOperation";
-export * from "./types/OperationPluginStatic";
-export * from "./types/OperationPluginView";
-export * from "./types/OperationPluginViewInfo";
+export * from "./domain/LocalTickModeString";
+export * from "./domain/OperationPlugin";
+export * from "./domain/OperationPluginInfo";
+export * from "./domain/OperationPluginOperation";
+export * from "./domain/OperationPluginStatic";
+export * from "./domain/OperationPluginView";
+export * from "./domain/OperationPluginViewInfo";
 export * from "./pdi-types/PlatformPointEvent";
-export * from "./types/Player";
+export * from "./domain/Player";
 export * from "./pdi-types/ShaderUniform";
-export * from "./types/TextAlign";
-export * from "./types/TextAlignString";
-export * from "./types/TextBaseline";
-export * from "./types/TextMetrics";
-export * from "./types/TickGenerationModeString";
+export * from "./domain/TextAlign";
+export * from "./domain/TextAlignString";
+export * from "./domain/TextBaseline";
+export * from "./domain/TextMetrics";
+export * from "./domain/TickGenerationModeString";
