+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from "./index.common";
-export * from "./Game";
-export * from "./GameHandlerSet"; // NOTE: コンテンツから参照する必要はない
+export * from "./domain/Game";
+export * from "./domain/GameHandlerSet"; // NOTE: コンテンツから参照する必要はない
